requirements.txt
alembic==1.5.7
asn1crypto==1.4.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-DotEnv==0.1.2
flask-file-upload==0.1.5
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.5.1
fonttools==4.21.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
oscrypto==1.2.1
Pillow==8.1.2
pycodestyle==2.7.0
pycparser==2.20
pyHanko==0.4.0
pyhanko-certvalidator==0.12.1
python-barcode==0.13.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-pkcs11==0.7.0
pytz==2021.1
PyYAML==5.4.1
qrcode==6.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.2
toml==0.10.2
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
